requirements.txt
django==5.2.1
django-cors-headers==4.9.0
openai==2.28.0
python-dotenv==1.0.0